backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
supabase==2.4.5
pdfplumber==0.11.0
pandas==2.2.2
openpyxl==3.1.2
google-genai
qdrant-client==1.9.1
python-docx==1.1.2
reportlab==4.2.0
